+++ tests/unit/Fixtures/ProductImage.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LAG\AdminBundle\Tests\Unit\Fixtures;
+
+use LAG\AdminBundle\Entity\Image;
+
+class ProductImage extends Image
+{
+}